\name{suttonbm-Rutils-package}
\alias{suttonbm-Rutils-package}
\alias{suttonbm-Rutils}
\docType{package}
\title{
\packageTitle{suttonbm-Rutils}
}
\description{
\packageDescription{suttonbm-Rutils}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{suttonbm-Rutils}
\packageIndices{suttonbm-Rutils}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{suttonbm-Rutils}

Maintainer: \packageMaintainer{suttonbm-Rutils}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
